IT'S ALIVE!!!!
This project is pretty incredible. This is at least a few weeks work right here. Now I can index the site, format the ads for this project, output to csv, put it in massive scale into dynamodb, and start to see what's out there. Out.csv are the results from the sample file. Everything is fully working. It's rough, but I'll be doing major refactorings through the week. This took all night to get running.
The results are very interesting. You have a list of all the persons and there is even a price, AGE, height, cup size, location, ethnicity. Even the restrictions are listed: ethnicity and age. It's pretty bad when you think about it. Really bad.

This image is fitting for this context - Slave Market in Rome - https://www.the-athenaeum.org/art/detail.php?ID=6037

Have humans evolved over the last 2,000 years? Will we ever? Or will people always be bought and sold for their flesh in one way or another?