Major refactorings and unit tests. Note: all changes to the output were caused by modifying int age_end = body.indexOf(",>           "); TO int age_end = body.indexOf(",>");. Other than the noted change all the refactorings were verified and continuously tested by the sample input and output files in source control.
Fixes:

 - Read input file as utf
 - Modified age target string to match the input, because age was stripped from most ad results

Description of changes:
I fixed the input file being read as ascii by specifying the encoding as UTF-8. I also discovered a major issue where age was not being included. It appears to be included from the "Posters age" section in the body, but age is on the header along with phone number and name for most ads. I'm not sure why that was wrong, but you can see in the csv that the quality of the results dramatically increased by filling in the -1 with valid ages, so I mean based on the existing input and output I've made improvements yet alone with all new incoming data. I archived all the current results in a unit test then added a new modification in those initial tests. Lastly I broke out a new test with new output from today. It worked well, the ad was very stripped down (although the link did remove the full path on the url it just shows as facebook). That's probably for crawling purposes and discovery for uniqueness. I'm not sure, but to me that's an issue. I can strip down to the domain in another step. I want the full url's on extraction.

Something is odd with age being stripped in the output file. I don't know what the purpose of the output file even is. It just seems like copy of the input for verification purposes, except that's not the case now with age being stripped. For example here is a sample row on the output now. I'm holding back the Out.txt changes so I remember to review. Need to sleep for the night, but I'm at the point where I trust this repo and my changes more than what was there.

BEFORE

???I'm ready for some fun! Come play!??? BLONDE - 23,> Hello

AFTER

???I'm ready for some fun! Come play!??? BLONDE

Tomorrow I'll figure out exactly how the input is being parsed and see if I can figure out what's going on with the output file and its purpose. The goal is to wire something up to .NET by the end of the week with .NET doing the crawling and being the "command center". It may be something like:

 - Crawl the current page every second
 - Parse the results for memex with the title and body with ads separated by a line break
 - Dump to a text file
 - Call the memex jar from .net and export to a csv
 - Convert the csv -> Plain Old Class Model (POCO) -> DynamodDb KVP

This would be really good progress. Then just scale out. Finally move onto new sites. Potentially crawl the links, do searches on phone numbers, crawl entirely new sites. Maybe even look into cralwers like nutch.